justfile: Add just as the command runner
`Just` is a nice alternative to make with programming languages that
already have their own build system, it acts as a command runner
instead. It's not necessary for iso-builders operation but it would be useful to
run the following in a CI environment:

just build && just run


Its syntax is basically a simplified and streamlined makefile, it uses
the justfile file name to declare a set of commands for `just` to run.
In the future it would be good to add a taskfile as well as we seem to
be introducing it into the DPU Operator.

Install anywhere (Linux, Macos, Windows) with the following (replace /usr/bin with wherever you
want the binary):
curl --proto '=https' --tlsv1.2 -sSf https://just.systems/install.sh | bash -s -- --to /usr/bin/