main.go: Add go implementation of our RHEL ISO Builder
The idea here is that you would run build this on an x86 machine, while
targetting arm64 (IPU's arch), which allows builds with no arm servers.
To do this build the binary in a container, and run that container with
the following flags:
sudo podman run --privileged --security-opt label=type:unconfined_t -it \
      --arch aarch64 \
      -v $PWD/workdir:/workdir \
      $IPU_BOOTC_BUILDER_URL \
      -u $BOOTC_IMAGE_URL

The `--priviledged` and `--security-opt label=type:unconfined_t` flags are specially important
to allow for multi arch which is enabled using the qemu-user-static
container on RHEL since it seems to have dropped qemu-user-static support in
RHEL 8

Another important thing to note is that this program interacts with
`/workdir` to do its function, so always at least mount a directory to
it.